requirements.txt
accelerate==1.2.1
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==25.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
biblioteca==0.0.1a2
bleach==6.1.0
blinker==1.9.0
blobfile==3.0.0
bokeh==3.6.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.1
certifi==2024.12.14
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.8
cloudpickle==3.0.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
community==1.0.0b1
contourpy==1.2.1
cryptography==43.0.1
cycler==0.12.1
dask==2024.9.1
dask-expr==1.1.15
dataclasses-json==0.6.7
datasets==3.3.0
debugpy==1.8.5
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.32.2
dill==0.3.8
distlib==0.3.9
distributed==2024.9.1
distro==1.9.0
durationpy==0.9
emoji==2.14.1
entrypoints==0.4
et-xmlfile==1.1.0
eval_type_backport==0.2.2
evaluate==0.4.3
executing==2.0.1
fabric==3.2.2
fake==0.8
Faker==29.0.0
fastapi==0.115.7
fastjsonschema==2.20.0
favicon==0.7.0
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
Flask==3.1.0
Flask-Cors==5.0.0
flatbuffers==25.1.24
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
ge==0.0.4
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.38.0
googleapis-common-protos==1.66.0
gritql==0.1.5
grpcio==1.70.0
h11==0.14.0
htbuilder==0.7.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.0
humanfriendly==10.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
invoke==2.2.0
ipykernel==6.29.5
ipython==8.26.0
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
json5==0.9.25
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
kubernetes==32.0.0
langchain==0.3.14
langchain-cli==0.0.35
langchain-community==0.3.14
langchain-core==0.3.33
langchain-openai==0.2.14
langchain-text-splitters==0.3.5
langdetect==1.0.9
langserve==0.3.1
langsmith==0.2.11
llama_models==0.0.61
llama_stack==0.0.61
llama_stack_client==0.0.61
locket==1.0.0
lxml==5.3.0
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.26.0
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
mmh3==5.1.0
modin==0.32.0
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
narwhals==1.14.2
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
ollama==0.4.4
onnxruntime==1.20.1
openai==1.77.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.15
overrides==7.7.0
packaging==24.2
pandas==2.2.2
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
partd==1.4.2
patsy==0.5.6
pdfminer.six==20231228
pdfplumber==0.11.5
pexpect==4.9.0
pillow==10.4.0
pip_search==0.0.13
platformdirs==4.2.2
plotly==5.24.1
plotly-express==0.4.1
posthog==3.11.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
propcache==0.2.1
protobuf==5.29.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyaml==24.12.1
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pymdown-extensions==10.12
PyMuPDF==1.25.3
PyMySQL==1.1.1
PyNaCl==1.5.0
pyparsing==3.1.2
pypdf==5.2.0
PyPDF2==3.0.1
pypdfium2==4.30.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-iso639==2025.1.28
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2024.1
pyvss==2024.6.0
PyYAML==6.0.2
pyzmq==26.1.1
RapidFuzz==3.12.1
ray==2.37.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.20.0
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.6.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
setuptools==75.8.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.37
sse-starlette==1.8.2
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
starlette==0.45.3
statsmodels==0.14.3
streamlit==1.40.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.5.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-lottie==0.0.5
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
sympy==1.13.1
tblib==3.0.0
tenacity==9.0.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tiktoken==0.8.0
tinycss2==1.3.0
tokenizers==0.21.0
toml==0.10.2
tomlkit==0.13.2
toolz==0.12.1
torch==2.6.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.47.0
typer==0.15.1
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unicorn==2.1.1
unstructured==0.16.20
unstructured-client==0.29.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
validators==0.34.0
virtualenv==20.28.0
voila==0.5.7
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.2
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xxhash==3.5.0
xyzservices==2024.9.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
zstandard==0.23.0
